--- src/selectors/protocolData.js
@@ -1,3 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-export const protocolData = state => state.protocol.config.data;
